turf-recommendation-backend/requirements.txt
pandas
scikit-surprise
numpy==1.23.5
# ...other dependencies...
